refactor(reports): move toServiceReport and extractAccomplishmentsOptimized to reports_helpers.go
Both functions are shared/pure helpers with no dependency on their
previous host files. Addresses PR #19 review feedback.
